Add ghostscript@9.27, librsvg@2.42.7 as externals for expanse/0.17.1/cpu
We've run into significant trouble building the lastest versions of
ghostscript natively in the expanse/0.17.1/cpu instance software stack
due to problems compiling rust, which is now a dependency for librsvg.
To work around this issue, we're currently using ghostscript from
Expanse's Rocky 8 operating system as an external dependency. However,
it would be more ideal to retry the spack-native ghostscript build using
only the librsvg library from Rocky as an external dependency, which is
also already available on Expanse, albeit an older version that likely
does not depend on rust. Another option may be to remove librsvg from
ghostscript as a dependnecy in the spack package itself, which should
have a minor impact to functionality at first glance.