Add vmd@1.9.3 % gcc@10.2.0 to expanse/0.17.1/cpu
A custom vmd spaclk package was added to the sdsc package repo to
correct the sha256 hash of the tarball currently being distributed from
the developers.

Also, please note, vmd originally deployed for expanse/0.15.4/cpu in
production as a core spack package. However, the latest version of the
vmd spack package in spack v0.17.1 now requires a significant number of
large dependencies be available either as external system packages or
natively installed by spack. In order to limit bloating the size of the
spack core with installing these additional dependencies, vmd was moved
to the gcc@10.2.0 dependency chain where these requirements can be
easily met already.