Add sirius@6.5.7 % gcc@10.2.0 ^openmpi@4.1.1 to expanse/0.17.1/cpu
Please note that this version of the sirius@6.5.7.sh build spec did not
complete successfully. Try again on next iteration. This commit is to
place it in the gcc@10.2.0 ^openmpi@4.1.1 dependency chain in a logical
location relative to its use as a dependency for downchain applications.
e.g., see cp2k