FINALLY! Start aocc@3.2.0 package dependency chain for expanse/0.17.1/cpu
The latest spack package for the AMD Optimizing C/C++ Compiler (AOCC) as
well as all of the latest spack packages for the AMD Optimizing CPU
Libraries (AOCL) were updated from the upstream spack/spack develop
branch and placed unmodified in the sdsc package repository. The main
difference(s) between the builtin AMD spack packages and the updated
ones in the sdsc repo is a minor version update from aocc 3.1 to 3.2 and
aocl 3.0 to 3.1.

There are a few build issues with the inital set of packages built using
aocc and the aocl libraries, namely, openblas@0.3.8-omp, arpack-ng@3.8.0,
and petsc@3.16.1. There are also appear to be spurious and/or unenforced
package dependencies. e.g., aocl-sparse indicates boost is a build and
link dependency, but never concretizes against boost. Addition open
questions remain from a support perspective. e.g., Should we always
attempt to use only the aocl libraries to satisfy package dependencies?
i.e., amdblis + amdlibflame to satisfy blas + lapack virtual
dependencies instead of openblas, which is currently used as the core
blas + lapack library for the gcc and intel compiler package dependency
chains. For the inital deployment of the aocc-built packages, we will
likely standardize dependencies on the same ones used for other compiler
chains, unless AMD has already tested use of the aocl libraries with a
given application and/or library.