Pause upgrade to spack v0.18.0
Initial test build of gcc@10.2.0 package dependency chain after
upgrading to spack v0.18.0 ran into a problem with some of the basic
standard Python packages [1]. It is unclear if there is a simple fix for
this issue at this time. As such, in order to move ahead with finalizing
a update of software stack for Expanse sooner rather than later, we will
pause the upgrade to spack v0.18.0 and continue finalizing the new
software stack on spack v0.17.2. We will reevaluate spack v0.18.x once
the update is complete.

[1]

==> Error: Class constructor failed for package 'builtin.py-virtualenv'.

Caused by:
AssertionError:
  File "/home/mkandes/cm/shared/apps/spack/0.18.0/cpu/lib/spack/spack/repo.py", line 1131, in get
    return package_class(spec)
  File "/home/mkandes/cm/shared/apps/spack/0.18.0/cpu/lib/spack/spack/package.py", line 798, in __init__
    spack.repo.get(self.extendee_spec)._check_extendable()
  File "/home/mkandes/cm/shared/apps/spack/0.18.0/cpu/lib/spack/spack/package.py", line 1350, in extendee_spec
    assert len(deps) == 1

ERROR: spack concretization failed.