.gitignore all *.exp-* standard output files
These are the standard output files generated during the deployment of
a spack instance on Expanse @ SDSC when using our custom slurm-based
spec build and install scripts. We may consider tracking these output
files within the sdsc/spack repository in the future. However, for the
time being, they are not necessarily needed to move forward with the new
deployment process.