Update for expanse/0.17.2/gpu instance build out
Please note that not all pakcages included here are passing build tests
yet. Some will also not be included in the final production instance to
be deployed. Other issues remain to be solved for some packages that
will eventually be included. e.g., verbs + cuda build problem for namd,
which can only be solved with updates to Spack package.py files for
charmpp and namd. Another complete rebuild pass through is planned next.