Revert relion package.py file to original version in spack v0.17.2
The latest version of the relion package.py file was tested with both
relion v3.1.4 and v4.0.0 on Expanse. However, the new libpng dependency
caused the builds, which used to succeed without issue, to fail. This
will need to be investigated in the future.

Note, however, the original builtins versions was modified to include
the latest relion versions and placed in the sdsc package repository.