Add ^cuda@11.2.2 to gromacs@2020.4 % intel@19.1.3.304 ^ intel-mpi@2019.10.317
Since intel-mpi does not have an explicit cuda dependency like openmpi,
we must to enforce the cuda version explictly when building against
intep-mpi. Otherwise, the default version of cuda will be installed,
which may or may not be the primary version deployed to support the
majority of the packages in the expanse/0.17.2/gpu spack instance.