Update all packages and dependency chains for expanse/0.17.3/gpu
The updated packages and dependency chains included in this commit were
used to deploy the new candidate production spack instance
expanse/0.17.3/gpu/a on Expanse in the /cm/shared Ceph filesystem.

Also added the new SPECS.md file with the aim to help improve
documentation of the package dependency chains.