Set ~knem ~xpmem in ucx@1.10.1 % gcc@8.5.0 and ucx@1.10.1 % gcc@10.2.0
During the update to spack v0.17.3, +knem and +xpmem were reset for ucx
to more closely match the AMD spack spec for openmpi [1]. However, it
was found that one or both variants when enabled in the latest deployed
pre-production test instances led to reduced performance for several
applications, namd with a charmpp backend=mpi in particular. In the next
iteration of the spack stack, which will be tested on devexp soon, we
will revisit this issue and attempt to set knem and xpmem as external
packages coming directly from the Mellanox OFED distribution installed
by the HPC Systems Group.

[1] https://developer.amd.com/spack/hpc-applications-openmpi