Remove arch-specific specs for externals in packages.yaml
The aim here is to simplify the structure of the packages.yaml file.
There are currently appears to be no advantages in specifying the
supported architecture(s) when referring to external packages. However,
we have observed Spack concretization and build issues when they are not
specified correctly and/or maintained properly. Hence, this simplification
to a non-architecture-specific definition of external packages in the
packages.yaml file should eliminate these problems and, hopefully, not
create any new ones.