rm mvapich2@2.3.7 % intel@19.1.3.304 from expanse/0.17.3/cpu
It is believed that either mvapich2@2.3.7 % intel@19.1.3.304 and/or one
of the MPI-based packages in its dependency chain are responsible for
triggering the loss of the lmod hierarchy during deployment we've now
observed multiple times on multiple systems [1]. Note, however, this
issue may be limited to Spack v0.17.x. As such, we may reconsider the
combined use of mvapich2 and intel compilers in the future for the
latest versions of Spack.

[1] https://github.com/spack/spack/issues/33897