Initial deployment of expanse/0.17.3/cpu/b is complete
This new Spack instance was built on and deployed into the /cm/shared
filesystem on Expanse from one of its AMD-based (Rome) compute nodes
with the latest updates to its system software:

  - Update operating system from Rocky 8.5 to Rocky 8.7
  - Update Infiniband driver from Mellanox OFED 4.9-4.1.7 to
    Mellanox OFED 5.8-1.1.2
  - Update Lustre client from lfs 2.12.8_6_g5457c37 to lfs 2.15.2
  - Update NVIDIA GPU driver from NVIDIA 510.39.01 CUDA 11.6 to
    NVIDIA 515.65.01 CUDA 11.7

These changes will be made system-wide on Monday, May 8th as part of
a planned maintenance. Functionality and performance testing of the
new instance will occur between now and then when it enters production
service for end users. Please note, however, the existing production
spack instance, expanse/0.15.4/cpu, is expected to remain in service
for some time following this upgrade.