rm lmod@8.5.6 % gcc@8.5.0 core expanse/0.17.3/gpu packages
On systems like Expanse or TSSC, we rely on NVIDIA's Bright Cluster
Manager to install a host OS and other software packages that will
serve Spack as external packages. In this case, lmod is provided by
Bright and is not needed (internally) for our Spack deployments on
SDSC's HPC systems.