Add wannier90@3.1.0 % gcc@10.2.0 ^openmpi@4.1.3 to expanse/0.17.3/cpu/b
Also included is an updated package.py file for wannier90 in the custom
sdsc package repo. This updated version is based on wannier90 package.py
file available in the spack/spack upstream on the date of this commite.
This fixes a build issue encountered with the builtins version available
in Spack v0.17.3. The fix was provided in this commit [1].

[1] https://github.com/spack/spack/commit/8332a5919483ed82cf65ba6f541b85ece2a75322