Add gromacs@2022.6 % gcc@10.2.0 ^openmpi@4.1.3 to expanse/0.17.3/gpu/b
A user requested an updated version of GROMACS for their work [1][2][3].
We decided it may be useful to deploy one for all users on the system in
the expanse/0.17.3/gpu/b production instance. At the time of this commit,
GROMACS v2022.6 was the latest release from the 2022.x version series,
which was released on Tuesday, July 11th, 2023 [4].

The build was tested successfully via my shared Spack instance
configuration for expanse/0.17.3/gpu/b. And the standard water benchmark
performance was observed within expected ns/day ranges. We will likely
deploy the same version in expanse/0.17.3/cpu/b as well to mirror the
instance configruations and options as closely as possible.

[1] https://github.com/sdsc/spack/issues/100
[2] https://sdsc.zendesk.com/agent/tickets/29944
[3] https://sdsc.zendesk.com/agent/tickets/26289
[4] https://manual.gromacs.org