Update all spec build scripts for the initial TSCC2 deployment
All of these spec build scripts are from the initial tscc/0.17.3/cpu
and tscc/0.17.3/gpu instances deployed on TSCC2 by Jerry Greenberg.
This commit is to document the changes made as part of the production
deployment and to synchronize them with the sdsc/spack repository.

Please note, however, this initial deployment came from an indirect
fork of the sdsc/spack repo [1]. With this commit, I will attempt to
force push the changes back to my fork of the sdsc/spack repo [2] with
the deployment branch renamed [3]. I will then attempt to create a pull
request from this branch in my mkandes/spack fork to the main
sdsc-0.17.3 deployment branch in sdsc/spack. If successful, I will
then attempt to commit the standard output files from the spec build
jobs to complete the documentation of the initial TSCC2 deployment.

[1] https://github.com/jerrypgreenberg/sdsc-tscc
[2] https://github.com/mkandes/spack
[3] git branch -m 0.17.3 tscc-0.17.3