Replicate blas and lapack % gcc@10.2.0 pkgs to % aocc@3.2.0 in exp/0.17.3/cpu/dev
Note, however, both openblas@0.3.18 threads=openmp and ucx@1.10.1 do not build
successfully with aocc@3.2.0. In the currently deployed production instance
exp/0.17.3/cpu/b, openblas@0.3.18 % aocc@3.2.0 is only offerred with threads=none
and the ucx@1.10.1 dependency for openmpi@4.1.3 is satisfied by compiling it with
the operating system's default compiler gcc@8.5.0. In the further development of
this initial development instance, we will forego the openmpi@4.1.3 ^ucx@1.10.1
issue and see if this resolves itself with we update to Spack v0.21.2. If not,
then we will reintroduce the packages dependent on openmpi with the same
split-compiler fix used in production. For now, we will attempt to
replicate the same aocc@3.2.0 package builds with mvapich2.3.7.