Set ~openmp for elpa@2021.05.001 % gcc@10.2.0 +mpi ^netlib-scalapack@2.1.0 ^openmpi@4.1.3
ELPA is currently intended to be used as a dependency for quantum-espresso@7.0,
which will also be a part of this development instance. Unfortuantely,
however, it was discovered that QE imposes the constraint of ~openmp on
the ELPA variant. As such, we will remove the default +openmp setting for
ELPA spec build scripts and installations, unless it is required in the
future to satisfy another package dependency.