Add zoltan@3.83 % aocc@3.2.0 ~fortran +parmetis ^mvapich2@2.3.7 to exp/0.17.3/cpu/dev
Like many other applications and/or libraries that depend on fortran,
their spec builds fail when compiled with aocc as their spack compiler.
As such, we've set ~fortran for this variant of zoltan. We can revisit
fortran support in the latest version of aocc when we upgrade to
Spack v0.21.2.