fftw@3.3.10 % gcc@10.2.0 +mpi +openmp failed builds
Unfortauntely, I have not been able to resolve this issue yet. It's a
bit odd that enabling +openmp causes the package build to fail when we
already have successful builds with +mpi ~openmp. We will need to
investigate this further in the future. For the time being, I will
invetigate alternative compiler, MPI, and FFTW options for QE.