Do not redirect spack-stage $TMPDIR (or $TMP) during bootstrap
While installing the first Spack package (bzip2@1.0.8 % gcc@8.5.0) for
the inew exp/0.17.3.gpu/dev instance via deploy.sh, it was observed that
the clingo bootstrapping process would fail when it attempted to utilize
a spack-stage directory in /tmp [1], which did not exist because we
generally always attempt to redirect the spack-stage directory to the
local /scratch disk on Expanse with in each Spack package spec build job
script. Here, I have commented out those redirections of TMPDIR and TMP
to avoid the issue in the future. If this problem is still present in the
latest versions of Spack, then we may want raise the issue with the Spack
developers as there should probably be no assumption on spack-stage
existing in /tmp for the bootstrapping process.

[1]

Input spec
--------------------------------
[    ]  .bzip2@1.0.8%gcc@8.5.0~debug+pic+shared

Concretized
--------------------------------
==> Bootstrapping clingo from pre-built binaries
==> Bootstrapping clingo from sources
[+] /usr (external bison-3.0.4-ihgvhux6mkzxcukac2op544m53rfl3rx)
==> Installing pkgconf-1.8.0-nodnfqwsxuesvfr2qjzqbueimsnd2pkb
==> No binary for pkgconf-1.8.0-nodnfqwsxuesvfr2qjzqbueimsnd2pkb found: installing from source
==> Fetching https://mirror.spack.io/_source-cache/archive/ef/ef9c7e61822b7cb8356e6e9e1dca58d9556f3200d78acab35e4347e9d4c2bbaf.tar.xz
==> Error: FileNotFoundError: [Errno 2] No such file or directory: '/tmp/mkandes/spack-stage/spack-stage-pkgconf-1.8.0-nodnfqwsxuesvfr2qjzqbueimsnd2pkb/pkgconf-1.8.0.tar.xz'