Add initial draft deploy.sh and all spec build scripts for exp/0.17.3/gpu/dev
This initial deploy.sh script and all of the supporting spec build scripts
for each Spack package now included as part of exp/0.17.3/gpu/dev has been
individually tested. This commit places this new development instance in a
state ready for a full deployment test. If this deployment test is a success,
then this version of exp/0.17.3/gpu/dev as-is will serve as the initial
gpu/dev instance. In addition, exp/0.17.3/cpu/dev deploy.sh will also be
reworked to align better with the pkg deployment graph of exp/0.17.3/gpu/dev.