Add qos=gpu-unlim and sleep breakpoints to deploy.sh for exp/0.17.3/gpu/dev
The initial full-scale deployment test of deploy.sh for exp/0.17.3/gpu/dev
failed due exceeding the MaxSubmitJobsPerAccount limit of the
qos=ind-gpu-shared-normal applied to partition=ind-gpu-shared on Expanse [1].
Here, I add qos=gpu-unlim to increase the maximum number of submitted
(and/or running) jobs to 52 and 15-minute sleep breakpoints to deploy.sh
to manually throttle the spec build job deployment at this time. Note,
however, this is likely not the best solution long-term, but it is a
temporary fix while a better solution is found. We may also need to
change the SLURM_CONF on Expanse to adjust the definition of qos=gpu-unlim
and authorize its use on partition=ind-gpu and ind-gpu-shared.

[1]

sbatch: error: MaxSubmitJobsPerAccount
sbatch: error: Batch job submission failed: Job violates accounting/QOS
               policy (job submit limit, user's size and/or time limits)