Uncomment #SBATCH --qos=gpu-unlim; add Slurm dep to subversion@1.14.0
qos=gpu-unlim should now be an AllowQos for the partition=ind-gpu-shared
in the SLURM_CONF for Expanse. This will be confirmed during the next
deployment test. Also included gdb@11.1 as a Slurm job dependncy for
subversion@1.14.0 in the core package layer in an attempt to avoid
multiple versions of python being installed implicitly, which was
observed in the recent exp/0.17.3/cpu/dev deployment testing due to
gdb's requirement that +debug be set for python.