Add beast2@2.6.4 % gcc@10.2.0 to deploy.sh for exp/0.17.3/gpu/dev
While libbeagle@3.1.2 % gcc@10.2.0 ^cuda@11.2.2 was added as part of the
exp/0.17.3/gpu/dev deployment job graph and the beast2@2.6.4 spec build
script already existed, it was not included as a part of the original
job graph constructed by deploy.sh for exp/0.17.3/gpu/dev. This correct
will be made for deploy.sh as well for exp/0.17.3/cpu/dev.