Add amdfftw@4.2 % gcc@13.2.0 ~openmp to exp/cpu/dev
Note, however, amdfftw@4.2 % gcc@13.3.0 +openmp failed to concretize
with the following error [1]. It is unclear how to resolve this conflict
at this time and, unfortunately, the use of --show-cores=full and
--show-cores=minimized options for debugging such conflicts with the
concretize appear to have been removed in the latest versions of Spack [2],
including v0.21.2. It is assumed the new error messaging model should be
sufficient without access to these options from clingo. I'm not sure this
was a good move. Why not offer the new error messaging model and the
ability to run the longer analysis, which has proven useful in the past
as well?

[1]

==> Error: concretization failed for the following reasons:
  1. 'amdfftw' required multiple values for single-valued variant 'openmp'

[2] https://github.com/spack/spack/pull/30669