Update boost@1.86.0 % gcc@13.3.0 for exp/cpu/dev
Here, I've switched on most of the boost libraries, which was how we
deployed boost previously for expanse/0.17.3/cpu/b and expanse/0.17.3/gpu/b.
The only potential issue here may be the incompatability with py-numpy@2.x.y.
I do not recall if this boost integration is a downstream package
dependency, e.g., like some of the other boost libraries required for cgal.