Update quantum-espresso@7.3.1 % gcc@13.3.0 +mpi ~openmp ^mvapich2@2.3.7
For some reason, QE requires git with +tcltk support to build correctly.
At least, that is how it appears after the first build tests here.
Otherwise, it fails on at some point on a ninja-related build error. As
such, I've set +tcltk as the default in the sdsc git package.py file to
avoid multiple installations of git within the different compiler layers
of the deployed instances.