Build ior@4.0.0 % gcc@13.3.0 ^mvapich2@2.3.7-2 fails due to autoconf
There remains a build problem with ior@4.0.0 for all compiler and MPI
combinations. I've placed the specific autoconf failure message within
the spec build script for ior@4.0.0 % gcc@13.3.0 ^mvapich2@2.3.7-2. We
will need to revisit and resolve the issue soon to restart filesystem
testing. A temp alternative would be to continue with ior@3.3.0.