Update deploy.sh for exp/cpu/dev
This is the first complete draft for the exp/cpu/dev instance based
on Spack v0.21.2. There remain a few packages with build issues. e.g.,
quantum-espresso, wps. However, the vast majority of the packages and
their spec build scripts run successfully. We will attempt to sort out
the remaining build issues as we move to deployment and performance
testing.