Update deploy.sh for exp/cpu/dev; fix initial AOCC dependency
Updated deploy.sh for exp/cpu/dev to fix the initial AOCC package
dependency chain to begin running once the compiler is ready to use. the
initial packages were (mistakenly) set to depend on GCC.

This update also include the addition of the reservation to the #SBATCH
parameters and the modified SPACK_INSTANCE_DIR to reflect the ongoing
deployment of exp/cpu/dev into the standard /cm/shared/apps/spack/0.21.2
production deployment path on Expanse.